+++ src/server/lib/mqtt/topics.js
@@ -1,0 +1,10 @@
+const topics = {
+  _report: 'report',
+  heater: {
+    stat: `stat/sonoff-heater/POWER`,
+    cmnd: `cmnd/sonoff-heater/power`,
+    sensor: `tele/sonoff-heater/SENSOR`,
+  }
+};
+
+module.exports = topics;
