+++ src/server/lib/routers/main.js
@@ -1,0 +1,9 @@
+const { Router } = require('express');
+
+const main = new Router();
+
+main.get('/', (req, res) => {
+  res.render('index');
+});
+
+module.exports = main;
